requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
httpx-socks==0.8.0
idna==3.7
motor==3.4.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pymongo==4.7.3
python-dotenv==0.21.1
python-socks==2.4.4
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.12.2
uvicorn==0.20.0
